Refactor matrix aggregation plumbing
Moves activity point reconstruction into `ActivityPoint+Matrix` and removes the old test-only `RecordReconstruction` helpers. `InMemoryDatabase` now adopts `MatrixAggregator` (instead of `ClientAggregating`), and stub databases no longer provide bespoke `activity`/`metricSeries` implementations. It also drops the no-op `CKDatabase` aggregation extension so aggregation behavior is defined through the shared matrix-based path.
