+++ Tests/ScoutTests/Transient/InMemoryDatabase.swift
@@ -56,17 +56,9 @@ final class InMemoryDatabase: DatabaseReader, RecordWriter, @unchecked Sendable 
             cursor: nil
         )
     }
-
-    func activity(in range: Range<Date>) async throws -> [ActivityPoint] {
-        try await reconstructedActivity(in: range)
-    }
-
-    func metricSeries(category: String, values: String, in range: Range<Date>) async throws -> [MetricSeries] {
-        try await reconstructedMetricSeries(category: category, values: values, in: range)
-    }
 }
 
-extension InMemoryDatabase: ClientAggregating {}
+extension InMemoryDatabase: MatrixAggregator {}
 
 extension InMemoryDatabase {
     var events: [Record] {
